TypeArrays

Documentation for TypeArrays.

TypeArrays.TypeArray

TypeArrays.TypeArray — Type

TypeArray <: StaticArray

An array that stores data entirely in the type.

This can be useful for e.g. very fast findfirst when lookup values are also known at compile time.